web/backend/requirements.txt
# Backend dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-socketio==5.11.0
python-multipart==0.0.9
pydantic==2.9.0
